+++ epyk/interfaces/components/CompLists.py
@@ -872,3 +872,28 @@ class Lists:
       self.page, items or [], width, height, html_code, helper, dflt_options, profile)
     html_f.input.attr['placeholder'] = placeholder
     return html_f
+
+  @html.Html.css_skin()
+  def menu(self, component, editable="fas fa-edit", refresh="fas fa-redo-alt", visible="fas fa-eye", height=(15, 'px'),
+           options=None, profile=None):
+
+    options = options or {}
+    menu_items = []
+    for typ, icon in [("Edit", editable), ("Add", visible), ("View", visible)]:
+      if icon:
+        r = self.page.ui.icons.awesome(
+          icon, text=typ, height=height, width=(35, 'px'), options=options, profile=profile)
+        r.span.style.css.line_height = r.style.css.height
+        r.icon.style.css.font_factor(-5)
+        r.style.css.font_factor(-5)
+        r.span.style.css.margin = "0 0 -3px -3px"
+        if typ == "Edit":
+          r.click([component.dom.setAttribute("contenteditable", True).r])
+        elif typ == "Add":
+          r.click([
+            component.dom.add("test")])
+        elif typ == "View":
+          r.click([component.dom.toggle()])
+        menu_items.append(r)
+    container = self.page.ui.div(menu_items, align="right", options=options, profile=profile)
+    return container