+++ CHANGELOG.md
@@ -4,16 +4,22 @@
 
 - Icons menu.
 - Add Bootstrap icon family from standard components.
+- Add paragraph as an Interface shortcut.
+- Add Texts menu.
+- Add Items lists menu.
+- Add copyToClipboard DOM shortcut to components.
 
 ### Changed
 
 - ChartJs JavaScript data() interface => align with build method.
 - ChartJs common series attributes.
+- Changed the Paragraph Markdown display
 
 ### Fixed
 
 - Fixed PivotTableJs component renderers.
-- Fixed Icon familiy use.
+- Fixed Icon family use.
+- Fixed color for content editable.
 
 ## [1.5.3] - 2021-03-12